api/requirements.txt
aiohttp==3.6.3
asyncinit==0.2.4
asyncpg==0.21.0
databases==0.4.1
fastapi==0.63.0
pre-commit
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart
requests==2.22.0
uvicorn==0.12.1
